# Convert lexer from Linked Linked list to Array based allocation