Benefits:

* Fewer allocations (1 per pool vs 1 per token)
* Better locality
* Smaller token struct size(no next pointer)
* Lots of functions that take pointers can take refs now instead
* Eliminate some tricky linked list code and replace it with clear range style code
* I think a similar design could be used for Loc objects (nodes/tokens would store a LocManager* and an index, 2 indices to have a LocRange)

Drawbacks

* Huge diff... not sure if that can be avoided, fortunately the parser is well covered by tests
* Implementation leaks info about the lexer.  Needed because ASM statements are parsed during semantic.  This could be improved if the parser was changed to operate on TokenRanges rather than call nextToken, etc.  This issue is also why Parser instances were changed from scope to auto.  Again, making the rest of the parser TokenRange aware would avoid the need for this.
* Extra variables needed in a few places because ref functions can't accept temporaries
* The choice of Pool parameters were pulled out of thin air... would probably be good to have something more principled
* a couple of API names seem bit cumbersome (makeRangeFromHere)