@rikkimax 2019-12-11
Can you please split out the memory ownership logic from the lexer class?
I.e. its own struct, that is never copied and is assumed to only have the one instance inside the lexer class.

Its out of scope of this PR, but it would be good to split out the tokenization from the lexer class as well. So that the class only acts as a accumulator of all the behavior and ties them together.