@thewilsonator 2019-12-11
> I think a similar design could be used for Loc objects

OMG, yes please! Doing that well would cut a massive amount of memory from the compiler.

> Huge diff... not sure if that can be avoided

Not so much of a problem. I'm sure reviewers will tell you how to break it up if need be.

> Implementation leaks info about the lexer. 

This is unfortunate. Since only ASM needs it, would it be possible to copy a block of tokens and make ASM use an array rather than a range of token.