@benjones 2019-12-11
I thought of giving the range a "detach" method that copies out the tokens or something, but the ASM parsing calls parseExpression, so I don't think it would work.