@benjones 2019-12-11
@rikkimax like a TokenPool struct?  I think each lexer would need its own, but yeah, that's an easy change.