@rikkimax 2019-12-11
@benjones Yeah I think you got the gist of what I was thinking about.

It would be nice to make that class a little simpler.
It may not be one of Walter's goals any time soon, but it would be nice to get that module ready to be split up.