@benjones 2019-12-11
Looks like the Windows_OMF x64 test is running out of memory.  Any suggestions on how to fix that?

Otherwise, I think I addressed the suggestions that don't need to go in a subsequent PR, and it's ready for a more thorough review