@WalterBright 2019-12-12
The DMD parser is highly sensitive to lexing speed. This needs a speed test before/after this change. I suggest using Phobos as the source data to use.