@benjones 2019-12-12
On my 5 year old MBP, it looks like this causes a slight slowdown on average.  I'll try to profile it and see if I can spot any obvious issues.  I assumed my division/modulo operations would be optimized to bitwise ops... does DMD do that?

Results for running make clean then time make (3 cycles, the first of which will rebuild druntime)
master:
27.50s, 17.67s, 17.66s
this PR with poolsize == 1024:
26.69s, 18.28s, 18.04s
This PR with poolsize = 256:
26.17s, 18.20s, 17.81s
This PR with poolsize = 64:
26.53s, 18.07s, 18.10s

PS = 256 seemed like a good tradeoff, although the speed doesn't seem to be too sensitive