@benjones 2019-12-12
Just using an Array is simpler code that's as fast or faster as the pool approach I was using.  It doesn't seem like anyone was relying on the address of a Token remaining constant