@wilzbach 2019-12-12
> On my 5 year old MBP, it looks like this causes a slight slowdown on average. I'll try to profile it and see if I can spot any obvious issues. I assumed my division/modulo operations would be optimized to bitwise ops...

I highly recommend only running your benchmark with LDC. DMD compiled with LDC is easily twice as fast (for math code it's even more), so I don't see why we should cater for DMD compiled with DMD for any kind of performance comparison and analysis.
In Phobos, for example, often refactorings result in slower code with DMD as it can't inline as much anymore, but we have a long list of precedents to accept them nevertheless if the performance with LDC doesn't degrade.