@benjones 2019-12-12
What's the easiest way to compile dmd with ldc?  `HOST_DMD=/my/ldc2 make -f posix.mak` gave me an error about the -vtls flag