@MoonlightSentinel 2019-12-12
@benjones Use the ldc wrapper `ldmd2`, it converts DMDs command line flags for ldc