@benjones 2019-12-12
timing make -f posix.mak in phobos after a clean, using dmd built with HOST_DMD=ldmd2, I get pretty much the same results.
First of each rebuilds druntime
master: 25.75, 17.83, 17.36
latest PR: 26.36, 18.10, 17.71

Running both in the profiler, very little time is spent in the lexer in either version