@MoonlightSentinel 2019-12-12
Did you use `ENABLE_RELEASE=1` and `ENABLE_LTO=1`?