@benjones 2019-12-12
Nope.  Take 3 :)

Results for building dmd with:  ENABLE_RELEASE=1 ENABLE_LTO=1 HOST_DMD=/usr/local/bin/ldmd2 VERBOSE=1 make -f posix.mak

master: 13.40, 9.64, 9.42
PR: 14.05, 9.56, 9.46

performance impact seems to be in the noise