@rainers 2019-12-14
Slightly offtopic: @CyberShadow can you please resurrect https://perf.dlang.io/ . Last results are from September.