@CyberShadow 2019-12-15
OK.

Looks like there is a compilation performance regression:

https://perf.dlang.io/#program-hello-compile-callgrind-insns;;1566740471.47766;1576430864

It will bisect towards it, should find the cause in a few hours.

(That, or something changed on the server OS installation to cause it; if it ends up showing 
the "cause" as the commit after f418f447f5fc (2019-09-19 06:36:06) then this would be the reason.)