@PetarKirov 2019-12-15
@benjones First off, thanks for working on this!
As you may have realized, doing an end-to-end build is not a good benchmark. Instead you should try to run just the just the lexer and parser on D source code.