@benjones 2019-12-16
@PetarKirov Any suggestions on how best to do that?  To some extent, the benchmarks show that this is more of a QOL improvement since very little of DMDs runtime is spent doing lexing AFAICT