@thewilsonator 2019-12-16
The thing about lexing is that it is on the critical path, meaning any slowdown will be realised 100% on all compilations.