@benjones 2019-12-16
@thewilsonator yeah, that makes sense.  Are there any lexing benchmarks around?  Is that something we want to add to the test suite before merging this PR?  

I haven't profiled DMD built with LDC, but for the one built with DMD, from what I could tell in the profiler, only about 100-200ms of the 17s phobos compile was spent in the lexer,