@benjones 2020-01-03
Now that the holidays are over I can keep working on this.... What should I do to make sure this doesn't introduce performance regressions?