@jacob-carlborg 2020-01-07
> Are there any lexing benchmarks around?

Here's a unit test for the diagnostic reporter for the lexer [1]. It shows an example how to only use the lexer.

[1] https://github.com/dlang/dmd/blob/master/test/unit/lexer/diagnostic_reporter.d