@benjones 2020-01-08
In trying to understand why my version was spending so much more time in malloc than master, I realized that I had fundamentally misunderstood how the current implementation works.   I had assumed that the entire token stream stays in memory as a linked list, when really only the currently "peeked" tokens live simultaneously and otherwise the current token just gets repeatedly overwritten...

Parts of this might still be useful, but I think this is probably no good in its current state.  I'll take a look at what's salvageable.